+++ Klassen/Schultag.cs
@@ -6,12 +6,15 @@ using System.Threading.Tasks;
 
 namespace Stundenplan.Klassen
 {
-    class Schultag
+    public class Schultag
     {
+        public string Wochentag { get; set; }
+
         public Stunde[] Stunden { get; set; }
 
-        public Schultag(Stunde[] Schulstunden)
+        public Schultag(string wochentag, Stunde[] Schulstunden)
         {
+            Wochentag = wochentag;
             Stunden = Schulstunden;
         }
     }
