+++ MainWindow.xaml
@@ -1,26 +1,69 @@
 ﻿<Window x:Class="Stundenplan.MainWindow"
+        x:Name="ThisWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:Klassen="clr-namespace:Stundenplan.Klassen"
         Title="Stundenplan">
-    <Grid Width="400">
-        <DataGrid x:Name="stundenplanData" CanUserSortColumns="False" CanUserReorderColumns="False" CanUserResizeColumns="False" AutoGenerateColumns="False" IsReadOnly="True" >
-            <DataGrid.Columns>
-                <DataGridTextColumn Header="Montag" Binding="{Binding Montag}">
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition/>
+        </Grid.RowDefinitions>
 
-                </DataGridTextColumn>
-                <DataGridTextColumn Header="Dienstag" Binding="{Binding Dienstag}">
+        <TextBlock HorizontalAlignment="Center" FontSize="20" Margin="0,20,0,20">Stundenplan</TextBlock>
+        <Button HorizontalAlignment="Right" Margin="0,20,10,20" Click="WochenansichtAendern">Wochenansicht wechseln</Button>
 
-                </DataGridTextColumn>
-                <DataGridTextColumn Header="Mittwoch" Binding="{Binding Mittwoch}">
+        <!--Tage-->
+        <ItemsControl Grid.Row="1"
+                      ItemsSource="{Binding ElementName=ThisWindow, Path=WochenplanData.Schultage}">
+            <ItemsControl.ItemsPanel>
+                <ItemsPanelTemplate>
+                        <UniformGrid Columns="{Binding ElementName=ThisWindow, Path=AnzahlSpalten}"/>
+                </ItemsPanelTemplate>
+            </ItemsControl.ItemsPanel>
 
-                </DataGridTextColumn>
-                <DataGridTextColumn Header="Donnerstag" Binding="{Binding Donnerstag}">
+            <ItemsControl.ItemTemplate>
+                <DataTemplate DataType="{x:Type Klassen:Schultag}">
+                    <Border BorderBrush="Black"
+                            BorderThickness="2">
+                        <HeaderedContentControl Header="{Binding Wochentag}"
+                                                HorizontalAlignment="Stretch">
+                            <HeaderedContentControl.HeaderTemplate>
+                                <DataTemplate>
+                                    <TextBlock Text="{Binding}"
+                                               Margin="5,0">
+                                        <TextBlock.TextDecorations>
+                                            <TextDecoration Location="Underline"
+                                                            PenOffset="4">
+                                                <TextDecoration.Pen>
+                                                    <Pen Thickness="3" Brush="Black"/>
+                                                </TextDecoration.Pen>
+                                            </TextDecoration>
+                                        </TextBlock.TextDecorations>
+                                    </TextBlock>
+                                </DataTemplate>
+                            </HeaderedContentControl.HeaderTemplate>
 
-                </DataGridTextColumn>
-                <DataGridTextColumn Header="Freitag" Binding="{Binding Freitag}">
-
-                </DataGridTextColumn>
-            </DataGrid.Columns>
-        </DataGrid>
+                            <!--Stunden-->
+                            <ItemsControl ItemsSource="{Binding Stunden}"
+                                          Margin="0,7,0,0">
+                                <ItemsControl.ItemTemplate>
+                                    <DataTemplate DataType="{x:Type Klassen:Stunde}">
+                                        <Button Click="Stunde_Click"
+                                                Margin="3"
+                                                Padding="3,1">
+                                            <Grid>
+                                                <TextBlock Text="{Binding Titel}"
+                                                           TextWrapping="Wrap" />
+                                            </Grid>
+                                        </Button>
+                                    </DataTemplate>
+                                </ItemsControl.ItemTemplate>
+                            </ItemsControl>
+                        </HeaderedContentControl>
+                    </Border>
+                </DataTemplate>
+            </ItemsControl.ItemTemplate>
+        </ItemsControl>
     </Grid>
 </Window>
