+++ MainWindow.xaml.cs
@@ -1,5 +1,7 @@
-﻿using System.Collections.ObjectModel;
+﻿using Stundenplan.Klassen;
+using System.Collections.Generic;
 using System.Windows;
+using System.Windows.Controls;
 
 namespace Stundenplan
 {
@@ -8,33 +10,85 @@ namespace Stundenplan
     /// </summary>
     public partial class MainWindow : Window
     {
-        private WochenplanReihe[] wochenplan;
-        public MainWindow()
+
+        public Wochenplan WochenplanData
         {
-            InitializeComponent();
-            wochenplan = new WochenplanReihe[]{
-                new WochenplanReihe("Deutsch","Musik", "Mathe", "Informatik","Englisch")
-            };
-            stundenplanData.ItemsSource = wochenplan;
+            get => (Wochenplan)GetValue(WochenplanDataProperty);
+            set => SetValue(WochenplanDataProperty, value);
         }
 
-    }
+        // Using a DependencyProperty as the backing store for AusgewaehlteStunde.  This enables animation, styling, binding, etc...
+        public static readonly DependencyProperty WochenplanDataProperty =
+            DependencyProperty.Register("WochenplanData", typeof(Wochenplan), typeof(MainWindow), new PropertyMetadata(null));
 
-    public class WochenplanReihe
-    {
-        public string Montag { get; set; }
-        public string Dienstag { get; set; }
-        public string Mittwoch { get; set; }
-        public string Donnerstag { get; set; }
-        public string Freitag { get; set; }
-
-        public WochenplanReihe(string montag, string dienstag, string mittwoch, string donnerstag, string freitag)
+        public int AnzahlSpalten
         {
-            Montag = montag;
-            Dienstag = dienstag;
-            Mittwoch = mittwoch;
-            Donnerstag = donnerstag;
-            Freitag = freitag;
+            get => (int)GetValue(AnzahlSpaltenProperty);
+            set => SetValue(AnzahlSpaltenProperty, value);
+        }
+
+        // Using a DependencyProperty as the backing store for AusgewaehlteStunde.  This enables animation, styling, binding, etc...
+        public static readonly DependencyProperty AnzahlSpaltenProperty =
+            DependencyProperty.Register("AnzahlSpalten", typeof(int), typeof(MainWindow), new PropertyMetadata(1));
+
+        public Stunde AusgewaehlteStunde
+        {
+            get => (Stunde)GetValue(AusgewaehlteStundeProperty);
+            set => SetValue(AusgewaehlteStundeProperty, value);
+        }
+
+        // Using a DependencyProperty as the backing store for AusgewaehlteStunde.  This enables animation, styling, binding, etc...
+        public static readonly DependencyProperty AusgewaehlteStundeProperty =
+            DependencyProperty.Register("AusgewaehlteStunde", typeof(Stunde), typeof(MainWindow), new PropertyMetadata(null));
+
+        public MainWindow()
+        {
+            Schultag montag = new Schultag("Montag", new Stunde[]
+            {
+                new Stunde("Deutsch"),
+                new Stunde("Mathe"),
+            });
+            Schultag dienstag = new Schultag("Dienstag", new Stunde[]
+            {
+                new Stunde("Sport"),
+            });
+            Schultag mittwoch = new Schultag("Mittwoch", new Stunde[]
+            {
+                new Stunde("Englisch"),
+            });
+            Schultag donnerstag = new Schultag("Donnerstag", new Stunde[]
+            {
+                new Stunde("Deutsch"),
+            });
+            Schultag freitag = new Schultag("Freitag", new Stunde[]
+            {
+                new Stunde("Deutsch"),
+            });
+
+            WochenplanData = new Wochenplan(
+                new List<Schultag>() 
+                {
+                    montag,
+                    dienstag, 
+                    mittwoch, 
+                    donnerstag, 
+                    freitag
+
+                });
+
+            InitializeComponent();
+        }
+
+        private void Stunde_Click(object sender, RoutedEventArgs e)
+        {
+            Button clickedButton = (Button)sender;
+            AusgewaehlteStunde = (Stunde)clickedButton.DataContext;
+        }
+
+        private void WochenansichtAendern(object sender, RoutedEventArgs e)
+        {
+            Button clickedButton = (Button) sender;
+            AnzahlSpalten = AnzahlSpalten == 1? 5: 1;
         }
     }
 }
