--- Stundenplan/Klassen/Stunde.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Stundenplan.Klassen
-{
-    public class Stunde
-    {
-        public string Titel { get; set; }
-
-        public Stunde(string titel)
-        {
-            Titel = titel;
-        }
-    }
-
-}
