--- Stundenplan/MainWindow.xaml
@@ -1,69 +1,0 @@
-﻿<Window x:Class="Stundenplan.MainWindow"
-        x:Name="ThisWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:Klassen="clr-namespace:Stundenplan.Klassen"
-        Title="Stundenplan">
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition/>
-        </Grid.RowDefinitions>
-
-        <TextBlock Text="{Binding ElementName=ThisWindow, Path=AusgewaehlteStunde.Titel}" />
-
-        <!--Tage-->
-        <ItemsControl Grid.Row="1"
-                      ItemsSource="{Binding ElementName=ThisWindow, Path=Wochenplan.Schultage}">
-            <ItemsControl.ItemsPanel>
-                <ItemsPanelTemplate>
-                    <UniformGrid Columns="5"
-                                 Grid.IsSharedSizeScope="True"/>
-                </ItemsPanelTemplate>
-            </ItemsControl.ItemsPanel>
-
-            <ItemsControl.ItemTemplate>
-                <DataTemplate DataType="{x:Type Klassen:Schultag}">
-                    <Border BorderBrush="Black"
-                            BorderThickness="2">
-                        <HeaderedContentControl Header="{Binding Wochentag}"
-                                                HorizontalAlignment="Stretch">
-                            <HeaderedContentControl.HeaderTemplate>
-                                <DataTemplate>
-                                    <TextBlock Text="{Binding}"
-                                               Margin="5,0">
-                                        <TextBlock.TextDecorations>
-                                            <TextDecoration Location="Underline"
-                                                            PenOffset="4">
-                                                <TextDecoration.Pen>
-                                                    <Pen Thickness="3" Brush="Black"/>
-                                                </TextDecoration.Pen>
-                                            </TextDecoration>
-                                        </TextBlock.TextDecorations>
-                                    </TextBlock>
-                                </DataTemplate>
-                            </HeaderedContentControl.HeaderTemplate>
-
-                            <!--Stunden-->
-                            <ItemsControl ItemsSource="{Binding Stunden}"
-                                          Margin="0,7,0,0">
-                                <ItemsControl.ItemTemplate>
-                                    <DataTemplate DataType="{x:Type Klassen:Stunde}">
-                                        <Button Click="Stunde_Click"
-                                                Margin="3"
-                                                Padding="3,1">
-                                            <Grid>
-                                                <TextBlock Text="{Binding Titel}"
-                                                           TextWrapping="Wrap" />
-                                            </Grid>
-                                        </Button>
-                                    </DataTemplate>
-                                </ItemsControl.ItemTemplate>
-                            </ItemsControl>
-                        </HeaderedContentControl>
-                    </Border>
-                </DataTemplate>
-            </ItemsControl.ItemTemplate>
-        </ItemsControl>
-    </Grid>
-</Window>
